[REF] pipeline -> base_etl
