+++ base_etl/__manifest__.py
@@ -1,0 +1,11 @@
+{
+    "name": "base_etl",
+    "summary": "Building blocks and conventions for file-based ETL integrations.",
+    "author": "Glo Networks",
+    "website": "https://github.com/GlodoUK/odoo-addons",
+    "category": "Technical",
+    "version": "19.0.1.0.0",
+    "depends": ["base"],
+    "external_dependencies": {"python": ["openpyxl", "xlrd", "xlwt"]},
+    "license": "LGPL-3",
+}
