+++ base_etl/tests/__init__.py
@@ -1,0 +1,6 @@
+from . import test_batch
+from . import test_codec_for
+from . import test_csv
+from . import test_files
+from . import test_xls
+from . import test_xlsx
