--- pipeline/__manifest__.py
@@ -1,11 +1,0 @@
-{
-    "name": "pipeline",
-    "summary": "Observable delayed method pipelines for Odoo models.",
-    "author": "Glo Networks",
-    "website": "https://github.com/GlodoUK/odoo-addons",
-    "category": "Pipeline",
-    "version": "19.0.3.0.0",
-    "depends": ["base"],
-    "external_dependencies": {"python": ["openpyxl", "xlrd", "xlwt"]},
-    "license": "LGPL-3",
-}
