--- pipeline/tests/__init__.py
@@ -1,7 +1,0 @@
-from . import test_batch
-from . import test_csv
-from . import test_files
-from . import test_pipeline
-from . import test_tools
-from . import test_xls
-from . import test_xlsx
