--- pipeline_queue_job/__manifest__.py
@@ -1,11 +1,0 @@
-{
-    "name": "pipeline_queue_job",
-    "summary": "Run pipeline delayed stages through OCA queue_job.",
-    "author": "Glo Networks",
-    "website": "https://github.com/GlodoUK/odoo-addons",
-    "category": "Pipeline",
-    "version": "19.0.1.0.0",
-    "depends": ["pipeline", "queue_job"],
-    "post_load": "post_load",
-    "license": "LGPL-3",
-}
