--- pipeline_queue_job/models/__init__.py
@@ -1,1 +1,0 @@
-from . import mixin
