--- pipeline_queue_job/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_with_delay
