+++ queue_job_trace/models/__init__.py
@@ -1,0 +1,2 @@
+from . import queue_job
+from . import base
