+++ queue_job_trace/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_trace
